Refs #5631 Moving TopicDataType
