+++ examples/C++/Benchmark/Benchmark_smallPubSubTypes.h
@@ -23,7 +23,7 @@
 #ifndef _BENCHMARK_SMALL_PUBSUBTYPES_H_
 #define _BENCHMARK_SMALL_PUBSUBTYPES_H_
 
-#include <fastrtps/TopicDataType.h>
+#include <fastrtps/topic/TopicDataType.h>
 
 #include "Benchmark_small.h"
