+++ examples/C++/DeadlineQoSExample/deadlinepayloadPubSubTypes.h
@@ -23,7 +23,7 @@
 #ifndef _DEADLINEPAYLOAD_PUBSUBTYPES_H_
 #define _DEADLINEPAYLOAD_PUBSUBTYPES_H_
 
-#include <fastrtps/TopicDataType.h>
+#include <fastrtps/topic/TopicDataType.h>
 
 #include "deadlinepayload.h"
