+++ examples/C++/Filtering/FilteringExamplePubSubTypes.h
@@ -23,7 +23,7 @@
 #ifndef _FILTERINGEXAMPLE_PUBSUBTYPES_H_
 #define _FILTERINGEXAMPLE_PUBSUBTYPES_H_
 
-#include <fastrtps/TopicDataType.h>
+#include <fastrtps/topic/TopicDataType.h>
