+++ examples/C++/FlowControlExample/FlowControlExamplePubSubTypes.h
@@ -23,7 +23,7 @@
 #ifndef _FLOWCONTROLEXAMPLE_PUBSUBTYPES_H_
 #define _FLOWCONTROLEXAMPLE_PUBSUBTYPES_H_
 
-#include <fastrtps/TopicDataType.h>
+#include <fastrtps/topic/TopicDataType.h>
 
 #include "FlowControlExample.h"
