+++ examples/C++/HelloWorldExampleTCP/HelloWorldPubSubTypes.h
@@ -23,7 +23,7 @@
 #ifndef _HELLOWORLD_PUBSUBTYPES_H_
 #define _HELLOWORLD_PUBSUBTYPES_H_
 
-#include <fastrtps/TopicDataType.h>
+#include <fastrtps/topic/TopicDataType.h>
