+++ examples/C++/XMLProfiles/XMLProfilesExamplePubSubTypes.h
@@ -23,7 +23,7 @@
 #ifndef _XMLPROFILESEXAMPLE_PUBSUBTYPES_H_
 #define _XMLPROFILESEXAMPLE_PUBSUBTYPES_H_
 
-#include <fastrtps/TopicDataType.h>
+#include <fastrtps/topic/TopicDataType.h>
