+++ fastrtpsgen/src/main/java/com/eprosima/fastrtps/idl/templates/RTPSPubSubTypeHeader_backup.stg
@@ -21,7 +21,7 @@ $fileHeader(ctx=ctx,  file=[ctx.filename, "PubSubTypes.h"], description=["This h
 #ifndef _$ctx.filename;format="toUpper"$_PUBSUBTYPES_H_
 #define _$ctx.filename;format="toUpper"$_PUBSUBTYPES_H_
 
-#include <fastrtps/TopicDataType.h>
+#include <fastrtps/topic/TopicDataType.h>
 
 using namespace eprosima::fastrtps;
