+++ include/fastrtps/transport/tcp/RTCPHeader.h
@@ -2,7 +2,7 @@
 #ifndef RTCP_HEADER_H
 #define RTCP_HEADER_H
 
-#include <fastrtps/TopicDataType.h>
+#include <fastrtps/topic/TopicDataType.h>
 #include <fastrtps/rtps/common/Types.h>
 #include <cstring>
 #include <fastcdr/FastCdr.h>
