+++ test/profiling/allocations/AllocTestTypePubSubTypes.h
@@ -23,7 +23,7 @@
 #ifndef _ALLOCTESTTYPE_PUBSUBTYPES_H_
 #define _ALLOCTESTTYPE_PUBSUBTYPES_H_
 
-#include <fastrtps/TopicDataType.h>
+#include <fastrtps/topic/TopicDataType.h>
 
 #include "AllocTestType.h"
